# openapi/oh.yaml
openapi: 3.1.0
info:
  title: Open Hospital API Documentation
  description: Open Hospital API Documentation
  contact:
    name: ApiInfo.DEFAULT_CONTACT
  license:
    name: GPL-3.0 license
    url: https://github.com/informatici/openhospital-api?tab=GPL-3.0-1-ov-file#readme
  version: 0.1.0
servers:
- url: http://API_HOST:API_PORT
security:
- bearerAuth: []
...
paths:
...
  /patients/{code}:
    get:
      tags:
      - Patients
      operationId: getPatient
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientDTO"
      security:
      - bearerAuth: []
    put:
      tags:
      - Patients
      operationId: updatePatient
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientDTO"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Patients
      operationId: deletePatient
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: boolean
      security:
      - bearerAuth: []
...
  /reports/patientexamrequest/{patientId}:
    get:
      tags:
      - Reports
      operationId: printPatientExamRequestPdf
      parameters:
      - name: patientId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                format: binary
      security:
      - bearerAuth: []
  /reports/patientexamination/{examinationId}:
    get:
      tags:
      - Reports
      operationId: printPatientExaminationPdf
      parameters:
      - name: examinationId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                format: binary
      security:
      - bearerAuth: []
  /reports/exams-list:
    get:
      tags:
      - Reports
      operationId: printExamsListPdf
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                format: byte
      security:
      - bearerAuth: []
  /reports/diseases-list:
    get:
      tags:
      - Reports
      operationId: printDiseasesListPdf
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                format: byte
      security:
      - bearerAuth: []
...
  /permissions:
    get:
      tags:
      - Permissions
      operationId: retrieveAllPermissions
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PermissionDTO"
      security:
      - bearerAuth: []
  /permissions/{id}:
    get:
      tags:
      - Permissions
      operationId: retrievePermissionById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionDTO"
      security:
      - bearerAuth: []
  /permissions/userGroupCode/{userGroupCode}:
    get:
      tags:
      - Permissions
      operationId: retrievePermissionsByUserGroupCode
      parameters:
      - name: userGroupCode
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PermissionDTO"
      security:
      - bearerAuth: []
  /permissions/name/{name}:
    get:
      tags:
      - Permissions
      operationId: retrievePermissionByName
      parameters:
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionDTO"
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    PatientDTO:
      type: object
      description: Class representing a patient
      properties:
        code:
          type: integer
          format: int32
          readOnly: true
        firstName:
          type: string
          description: First name of the patient
          example: Mario
          maxLength: 50
        secondName:
          type: string
          description: Last name of the patient
          example: Rossi
          maxLength: 50
        name:
          type: string
          readOnly: true
        birthDate:
          type: string
          description: Birth date
          example: 1979-05-01
        age:
          type: integer
          format: int32
          description: Age
          example: 40
        agetype:
          type: string
          description: Age type
          example: null
        sex:
          type: string
          description: Sex
          enum:
          - M
          - F
          example: M
        address:
          type: string
          description: Address
          example: "Via Roma, 12"
          maxLength: 50
        city:
          type: string
          description: City
          example: Verona
          maxLength: 50
        nextKin:
          type: string
          readOnly: true
        telephone:
          type: string
          description: Telephone
          example: "+393456789012"
          maxLength: 50
        note:
          type: string
          description: Note
          example: Test insert new patient
          maxLength: 65535
        motherName:
          type: string
          description: Mother's name
          example: Roberta
          maxLength: 50
        mother:
          type: string
          description: "Mother's status (D=dead, A=alive)"
          enum:
          - D
          - A
          example: A
        fatherName:
          type: string
          description: Father's name
          example: Giuseppe
          maxLength: 50
        father:
          type: string
          description: "Father's status (D=dead, A=alive)"
          enum:
          - D
          - A
          example: D
        bloodType:
          type: string
          description: "Blood type (0-/+, A-/+ , B-/+, AB-/+)"
          enum:
          - 0-
          - 0+
          - A-
          - A+
          - B-
          - B+
          - AB-
          - AB+
          example: A+
        hasInsurance:
          type: string
          description: "HasInsurance (Y=Yes, N=no)"
          enum:
          - "Y"
          - "N"
          example: "N"
        parentTogether:
          type: string
          description: "Parent together (Y=Yes, N=no)"
          enum:
          - "Y"
          - "N"
          example: "N"
        taxCode:
          type: string
          description: Tax code
          example: RSSMRA79E01L781N
          maxLength: 30
        lock:
          type: integer
          format: int32
          description: Lock
          example: 0
        blobPhoto:
          type: string
          format: byte
          description: BlobPhoto
        hashCode:
          type: integer
          format: int32
          readOnly: true
        allergies:
          type: string
          description: Allergies of patient
          maxLength: 255
        anamnesis:
          type: string
          description: Current anamnesis
          maxLength: 255
        status:
          type: string
          description: Status
          enum:
          - I
          - O
          example: I
        consensusFlag:
          type: boolean
          description: Consensus flag
          example: true
        consensusServiceFlag:
          type: boolean
          description: Consensus service flag
          example: true
      required:
      - age
      - bloodType
      - city
      - fatherName
      - firstName
      - motherName
      - secondName
      - sex
...
    PermissionDTO:
      type: object
      properties:
        name:
          type: string
        id:
          type: integer
          format: int32
          description: The permission id
        description:
          type: string
          description: The permission description
          maxLength: 255
      required:
      - description
      - id
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT